결제 준비시 payment history 생성 및 구독 결제시 보증금 계산 로직 추가 (이후 수정 보완 필요)

+++ src/main/java/com/imchobo/sayren_back/domain/subscribe/service/SubscribeServiceImpl.java
@@ -56,8 +56,10 @@ public class SubscribeServiceImpl implements SubscribeService {
       // 보증금 및 월 렌탈료 저장
       int monthlyFee = dto.getMonthlyFeeSnapshot();
       int depositSnapshot = calculateDeposit(monthlyFee);
-      subscribe.setMonthlyFeeSnapshot(monthlyFee);
-      subscribe.setDepositSnapshot(depositSnapshot);
+
+      // 스냅샷 값
+      subscribe.setMonthlyFeeSnapshot((long) monthlyFee);
+      subscribe.setDepositSnapshot((long) depositSnapshot);
 
       // 로그인 유저 주입
       Member currentMember = SecurityUtil.getMemberEntity(); // 또는 상위에서 받아온 member
